1d complete; tests pass; need to valgrind
git-svn-id: https://zxing.googlecode.com/svn/trunk@2606 59b500cc-1b3d-0410-9834-0bbf25fbcc57

+++ cpp/core/tests/src/qrcode/VersionTest.cpp
@@ -48,7 +48,7 @@ static void checkVersion(Version *version, int number, int dimension) {
 CPPUNIT_TEST_SUITE_REGISTRATION(VersionTest);
 
 void VersionTest::testVersionForNumber() {
-  for (size_t i = 1; i <= Version::VERSIONS.size(); i++) {
+  for (int i = 1; i <= (int)Version::VERSIONS.size(); i++) {
     Version *v = Version::VERSIONS[i-1];
     CPPUNIT_ASSERT_EQUAL((int)i, v->getVersionNumber());
   }
